+++ Server/src/main/java/com/IS442/teamsixtester/repository/VesselRepository.java
@@ -3,5 +3,7 @@ package com.IS442.teamsixtester.repository;
 import com.IS442.teamsixtester.domain.Vessel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VesselRepository extends JpaRepository<Vessel, Long> {
+import java.util.UUID;
+
+public interface VesselRepository extends JpaRepository<Vessel, Integer> {
 }
